test(cypress): fix move component test support file mini-store.js
