+++ test/cypress/support/component_test/mini-store.js
